+++ kernel/include/kernel/schedule.h
@@ -1,0 +1,15 @@
+#ifndef KERNEL_SCHEDULE_H_
+#define KERNEL_SCHEDULE_H_
+
+#include "kernel/task.h"
+
+#define TASK_SWITCH_FREQUENCY 20
+
+
+void init_schedule();
+
+
+
+
+
+#endif // KERNEL_SCHEDULE_H_